Stripe payment method APIs, Pagination and Search functionalities added

+++ user/serializers.py
@@ -5,7 +5,10 @@ from django.utils import timezone
 from django.contrib.auth import get_user_model, authenticate
 from rest_framework import serializers
 from .models import User, UserCode
-from datetime import timedelta
+from rest_framework import exceptions
+from django.core import mail
+import pyotp
+from datetime import timedelta, datetime
 
 
 class NewUserSerializer(serializers.ModelSerializer):
@@ -13,6 +16,24 @@ class NewUserSerializer(serializers.ModelSerializer):
     class Meta:
         model = get_user_model()
         fields = ['email', 'password', 'name', 'image', 'country', 'age']
+
+    def create(self, validated_data):
+        """Sending email and perform the user creation in the system"""
+        get_user_model().objects.filter(creation__lt=datetime.now() -
+                                        timedelta(minutes=3), tempUser=True).delete()
+        if get_user_model().objects.filter(email=validated_data['email']):
+            raise exceptions.ValidationError(
+                {"detail": "Another user with this email already exists."})
+        totp = pyotp.TOTP('base32secret3232')
+        code = totp.now()
+        with mail.get_connection() as connection:
+            mail.EmailMessage(
+                "Email Verification", f"<h1>${code}</h1>", "the-book-spot@admin.com", [
+                    validated_data['email']],
+                connection=connection,
+            ).send()
+        # serializer.save(user_code=code)
+        return get_user_model().objects.create_user(**validated_data, user_code=code)
 
 
 class UserSerializer(serializers.ModelSerializer):
@@ -30,21 +51,19 @@ class UserCodeSerializer(serializers.ModelSerializer):
 
     def create(self, validated_data):
         """Create a new user in the system"""
-
         tempUser = None
         try:
-            tempUser = User.objects.get(
+            tempUser = get_user_model().objects.get(
                 user_code=validated_data['user_code'])
         except:
             raise serializers.ValidationError(
                 {"detail": "Invalid OTP entered."})
-        if timezone.now() > (tempUser.creation+timedelta(minutes=3)):
+        if (timezone.now() > (tempUser.creation+timedelta(minutes=3)) and tempUser.tempUser):
             tempUser.delete()
             raise serializers.ValidationError(
                 {"detail": "Your verification code has been expired. Please try again!"})
-        tempUser.delete()
         tempUser.tempUser = False
-        tempUser.user_code = ''
+        tempUser.user_code = -1*tempUser.user_code
         tempUser.save()
         return super().create(validated_data)
 
@@ -63,7 +82,6 @@ class UpdateUserSerializer(serializers.ModelSerializer):
             if password:
                 user.set_password(password)
                 user.save()
-
             return user
 
 
@@ -86,7 +104,10 @@ class AuthTokenSerializer(serializers.Serializer):
         """Validate and authenticate the user"""
         email = attrs.get('email')
         password = attrs.get('password')
-
+        try:
+            user = get_user_model().objects.get(email=email, password=password)
+        except:
+            user = None
         user = authenticate(
             request=self.context.get('request'),
             username=email,
